feat(ia): email tab and section support
