+++ assets/wizards/newspack/views/settings/sections.tsx
@@ -8,12 +8,13 @@ import { __ } from '@wordpress/i18n';
 const settingsTabs = window.newspackSettings;
 
 import Connections from './connections';
+import Emails from './emails';
 
 type SectionKeys = keyof typeof settingsTabs;
 
 const sectionComponents: Record< SectionKeys | 'default', () => JSX.Element > = {
 	connections: Connections,
-	// emails: Emails,
+	emails: Emails,
 	// social: Social,
 	// syndication: Syndication,
 	// seo: Seo,
