+++ assets/wizards/types/wizard-action-card.d.ts
@@ -3,18 +3,20 @@
  */
 type ActionCardProps = Partial< {
 	title: string | JSX.Element;
+	titleLink?: string;
+	href?: string;
 	description: string | JSX.Element | ( () => JSX.Element | string );
 	actionText: JSX.Element | string | null;
 	badge: string;
 	className: string;
 	indent: string;
 	notification: string;
-	notificationLevel: 'error' | 'warning';
+	notificationLevel: 'error' | 'warning' | 'info';
 	isMedium: boolean;
 	disabled: boolean | string;
 	hasGreyHeader: boolean;
 	toggleChecked: boolean;
-	toggleOnChange: () => void;
+	toggleOnChange: ( a?: boolean ) => void;
 	actionContent: boolean | JSX.Element | null;
 	error: string | null;
 	handoff: string | null;
